Copy createQueryPreloader type tests over to signature tests
